retype of Stuart's work on main to testing and addition of Sizegrip element
git is giving me a headache after switching branches... it should run the same as main right now except we can mouse-over the element of the window that will eventually resize the windows on the bottom right corner (sg.Sizegrip). Stuart wrote all of this already in main branch in case github reads tricky, my original contributions are only line 27 and 61 for now.